third_party/blink/web_tests/http/tests/inspector-protocol/prerender/prerender-status-update-on-enable-expected.txt


Running test: testReady
Test that a status update is sent for a completed prerender on enable
{
    key : {
        action : Prerender
        loaderId : <string>
        url : http://127.0.0.1:8000/inspector-protocol/prerender/resources/empty.html
    }
    pipelineId : <string>
    status : Ready
}

Running test: testFailure
Test that a status update is sent for a completed prerender on enable
{
    key : {
        action : Prerender
        loaderId : <string>
        url : http://127.0.0.1:8000/echo?status=204
    }
    pipelineId : <string>
    prerenderStatus : PrerenderFailedDuringPrefetch
    status : Failure
}

Running test: testCandidateRemovedAfterFailure
Test that a status update is sent for a completed prerender on enable
No status update received after Preload domain is re-enabled.
